add AST to WordsSay; move WordsProgramExceptions to words.exceptions
